# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ITFY E-Voting Platform API
  description: |
    # ITFY E-Voting Platform Backend API
    
    Welcome to the comprehensive API documentation for the ITFY E-Voting platform.
    This platform enables event organizers to create and manage voting events with 
    candidates, categories, and secure payment-based voting.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **Event Management**: Create, manage, and publish voting events
    - **Category Management**: Organize candidates into voting categories
    - **Candidate Management**: Register and manage candidates with profiles
    - **Voting System**: Purchase vote bundles and cast votes securely
    - **Payment Integration**: Paystack payment gateway integration
    - **Forms & Submissions**: Dynamic nomination forms with approval workflows
    - **Analytics**: Comprehensive voting analytics and reporting
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **super_admin**: Full system access
    - **admin**: Administrative access to events and users
    - **organiser**: Can create and manage events
    - **moderator**: Can view and moderate content
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
...
    
    ## Pagination
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    CandidateAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated candidates
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "Detailed error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 95
        itemsPerPage:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: "507f1f77bcf86cd799439011"

    # ==================
    # User Schemas
    # ==================
...
    CandidateProject:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        url:
          type: string
        image:
          type: string
        date:
          type: string
          format: date-time
    CandidateEducation:
      type: object
      properties:
        institution:
          type: string
        qualification:
          type: string
        field:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
    CandidateSocialLinks:
      type: object
      properties:
        facebook:
          type: string
        twitter:
          type: string
        instagram:
          type: string
        linkedin:
          type: string
        youtube:
          type: string
        website:
          type: string
        tiktok:
          type: string
    Candidate:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        candidate_code:
          type: string
          description: Unique candidate code
        phone_number:
          type: string
        slug:
          type: string
        bio:
          type: string
        profile_image:
          type: string
          nullable: true
        cover_image:
          type: string
          nullable: true
        gallery:
          type: array
          items:
            type: string
        video_url:
          type: string
        projects:
          type: array
          items:
            $ref: '#/components/schemas/CandidateProject'
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            $ref: '#/components/schemas/CandidateEducation'
        social_links:
          $ref: '#/components/schemas/CandidateSocialLinks'
        status:
          type: string
          enum: [pending, approved, rejected, profile_update_pending, pending_profile_completion]
        event:
          $ref: '#/components/schemas/ObjectId'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        vote_count:
          type: integer
          minimum: 0
        is_featured:
          type: boolean
        is_published:
          type: boolean
        display_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCandidateRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - event
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone_number:
          type: string
        bio:
          type: string
        profile_image:
          type: string
        event:
          $ref: '#/components/schemas/ObjectId'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
...
    Form:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
          description: Form name
        description:
          type: string
          description: Form description
        slug:
          type: string
          description: URL-friendly identifier
        form_type:
          type: string
          enum: [nomination, registration, survey, feedback]
          description: Type of form
        event:
          $ref: '#/components/schemas/ObjectId'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
          description: Categories this form applies to
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
          description: Form field definitions
        settings:
          $ref: '#/components/schemas/FormSettings'
        status:
          type: string
          enum: [draft, active, closed, archived]
          default: draft
        submission_count:
          type: integer
          description: Total submissions received
        candidate_field_mapping:
          type: object
          description: Mapping of form fields to candidate fields
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FormSettings:
      type: object
      properties:
        allow_anonymous:
          type: boolean
          default: false
        require_email_verification:
          type: boolean
          default: false
        submissions_per_user:
          type: integer
          description: Max submissions per user
        submissions_per_ip:
          type: integer
          description: Max submissions per IP
        opens_at:
          type: string
          format: date-time
        closes_at:
          type: string
          format: date-time
        confirmation_message:
          type: string
        redirect_url:
          type: string
        duplicate_detection:
          type: object
          properties:
            enabled:
              type: boolean
            method:
              type: string
              enum: [exact_match, fuzzy_match, field_similarity, email_match, phone_match]
            threshold:
              type: number
              description: Similarity threshold (0-100)
            fields:
              type: array
              items:
                type: string
              description: Fields to check for duplicates
...
    Submission:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        form:
          oneOf:
            - $ref: '#/components/schemas/ObjectId'
            - $ref: '#/components/schemas/Form'
        event:
          $ref: '#/components/schemas/ObjectId'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        submission_data:
          type: object
          additionalProperties: true
          description: Form field values keyed by field_id
        normalized_data:
          type: object
          additionalProperties:
            type: string
          description: Normalized data for duplicate detection
        status:
          type: string
          enum: [pending, approved, rejected, duplicate, under_review]
          default: pending
        submitted_by:
          type: object
          properties:
            user_id:
              $ref: '#/components/schemas/ObjectId'
            name:
              type: string
            email:
              type: string
            phone:
              type: string
        ip_hash:
          type: string
          description: Hashed IP address
        user_agent:
          type: string
        session_id:
          type: string
        duplicate_check:
          type: object
          properties:
            is_duplicate:
              type: boolean
            similarity_score:
              type: number
              minimum: 0
              maximum: 100
            matched_submissions:
              type: array
              items:
                type: object
                properties:
                  submission_id:
                    $ref: '#/components/schemas/ObjectId'
                  similarity:
                    type: number
                  matched_fields:
                    type: array
                    items:
                      type: string
            checked_at:
              type: string
              format: date-time
        nominee_identifier:
          type: string
          description: Unique identifier for nominee across categories
        approval:
          type: object
          properties:
            reviewed_by:
              $ref: '#/components/schemas/ObjectId'
            reviewed_at:
              type: string
              format: date-time
            review_notes:
              type: string
            rejection_reason:
              type: string
        candidate:
          $ref: '#/components/schemas/ObjectId'
          description: Created candidate (if approved)
        attachments:
          type: array
          items:
            type: object
            properties:
              field_id:
                type: string
              filename:
                type: string
              url:
                type: string
              size:
                type: integer
              mime_type:
                type: string
              uploaded_at:
                type: string
                format: date-time
        submission_number:
          type: string
          description: Human-readable submission reference
        confirmation_sent:
          type: boolean
        confirmation_sent_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSubmissionRequest:
      type: object
      required:
        - form
        - submission_data
      properties:
        form:
          $ref: '#/components/schemas/ObjectId'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        submission_data:
          type: object
          additionalProperties: true
          description: Field values keyed by field_id
        submitted_by:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
...
    ApproveSubmissionRequest:
      type: object
      properties:
        review_notes:
          type: string
        create_candidate:
          type: boolean
          default: true
          description: Whether to create a candidate from this submission
        candidate_data:
          type: object
          properties:
            name:
              type: string
            code:
              type: string
            category:
              $ref: '#/components/schemas/ObjectId'
          description: Override data for candidate creation
    RejectSubmissionRequest:
      type: object
      required:
        - rejection_reason
      properties:
        rejection_reason:
          type: string
          minLength: 1
        review_notes:
          type: string
        send_notification:
          type: boolean
          default: true
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Forbidden"
            error: "Access denied. You don't have permission for this resource"
    NotFoundError:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Not Found"
            error: "The requested resource was not found"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation Error"
            error: "Some required fields are missing or invalid"
...
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page
    SearchParam:
      in: query
      name: search
      schema:
        type: string
      description: Search query string
...
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/ObjectId'
      description: Resource ID
...
paths:
...
  /candidates:
    get:
      tags:
        - Candidates
      summary: List all candidates
      description: Retrieve paginated list of candidates (Admin/Organiser only)
      operationId: listCandidates
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - in: query
          name: event
          schema:
            type: string
          description: Filter by event ID
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, profile_update_pending, pending_profile_completion]
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category ID
      responses:
        '200':
          description: Candidates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Candidate'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags:
        - Candidates
      summary: Create a new candidate
      description: Create a new candidate (Admin/Organiser only)
      operationId: createCandidate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCandidateRequest'
      responses:
        '201':
          description: Candidate created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Candidate'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /submissions:
    post:
      tags:
        - Submissions
      summary: Submit a form
      description: Submit a form response (Public or authenticated based on form settings)
      operationId: submitForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '201':
          description: Submission created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          submission:
                            $ref: '#/components/schemas/Submission'
                          submission_number:
                            type: string
                          message:
                            type: string
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - Submissions
      summary: List all submissions
      description: Get paginated list of submissions with filters (Admin only)
      operationId: listSubmissions
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: form
          in: query
          schema:
            type: string
          description: Filter by form ID
        - name: event
          in: query
          schema:
            type: string
          description: Filter by event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, duplicate, under_review]
        - name: is_duplicate
          in: query
          schema:
            type: boolean
          description: Filter by duplicate status
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /submissions/me:
    get:
      tags:
        - Submissions
      summary: Get my submissions
      description: Get current user's form submissions
      operationId: getMySubmissions
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: form
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, duplicate, under_review]
      responses:
        '200':
          description: User's submissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /submissions/form/{formId}/duplicates:
    get:
      tags:
        - Submissions
      summary: Get suspected duplicates for form
      description: Get submissions flagged as potential duplicates (Admin only)
      operationId: getSuspectedDuplicates
      security:
        - BearerAuth: []
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: min_similarity
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Minimum similarity score to include
      responses:
        '200':
          description: Suspected duplicates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /submissions/{id}/approve:
    post:
      tags:
        - Submissions
      summary: Approve submission
      description: Approve a submission and optionally create a candidate (Admin/Organiser only)
      operationId: approveSubmission
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveSubmissionRequest'
      responses:
        '200':
          description: Submission approved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          submission:
                            $ref: '#/components/schemas/Submission'
                          candidate:
                            $ref: '#/components/schemas/Candidate'
                            description: Created candidate (if create_candidate=true)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /submissions/{id}/reject:
    post:
      tags:
        - Submissions
      summary: Reject submission
      description: Reject a submission with reason (Admin/Organiser only)
      operationId: rejectSubmission
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectSubmissionRequest'
      responses:
        '200':
          description: Submission rejected
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Submission'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...